requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
PyMuPDF==1.24.7
python-docx==1.1.2
numpy==2.0.1
scikit-learn==1.5.1
scipy==1.14.0
google-generativeai==0.7.2
orjson==3.10.5
reportlab==4.2.2
pytest==8.3.2
requests==2.32.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
langchain-core==0.2.12
langchain-community==0.2.10
langchain-text-splitters==0.2.2
unstructured==0.14.9
pypdf==4.2.0
docx2txt==0.8.1
